+++ stdlib/InteractiveUtils/src/editless.jl
@@ -226,7 +226,12 @@ edit(m::Module) = edit(pathof(m))
 if Sys.iswindows()
     function less(file::AbstractString, line::Integer)
         pager = shell_split(get(ENV, "PAGER", "more"))
-        g = pager[1] == "more" ? "" : "g"
+        if pager[1] == "more"
+            g = ""
+            line -= 1
+        else
+            g = "g"
+        end
         run(Cmd(`$pager +$(line)$(g) \"$file\"`, windows_verbatim = true))
         nothing
     end